requirements.txt
fastapi==0.115.0
joblib==1.4.2
numpy==1.24.3
pandas==2.0.1
protobuf==5.28.2
google-cloud-storage==2.8.0
gunicorn
